refactor(web): split core/lib/chatActions/crossActions.ts into per-handler modules
